# RNET-1131: optimize many OR'd terms on UUID/ObjectId queries