--- a/src/realm/query_engine.hpp
+++ b/src/realm/query_engine.hpp
@@ -71,6 +71,7 @@ TConditionValue:    Type of values in condition column. That is, int64_t, float,
 #include <functional>
 #include <sstream>
 #include <string>
+#include <typeindex>
 
 #include <realm/array_backlink.hpp>
 #include <realm/array_basic.hpp>
@@ -226,7 +227,7 @@ public:
         return s;
     }
 
-    bool consume_condition(ParentNode& other, bool ignore_indexes)
+    bool consume_condition(ParentNode& other, bool ignore_indexes, std::optional<size_t> num_identical_conditions)
     {
         // We can only combine conditions if they're the same operator on the
         // same column and there's no additional conditions ANDed on
@@ -246,9 +247,20 @@ public:
         // choose 2. The exception is if we're inside a Not group or if the query is restricted to a view, as in those
         // cases end will always be start+1 and we'll have O(N*M) runtime even with a search index, so we want to
         // combine even with an index.
-        if (has_search_index() && !ignore_indexes)
+        //
+        // If the column is not a string type, then as the number of conditions increases, the cost of using the index for each condition
+        // rises faster than using a hash set to check if each value in a leaf matches a condition.
+        // So if there are _many_ conditions (as defined below), combine conditions even if an index is present.
+        if (has_search_index() && !ignore_indexes &&
+            (m_condition_column_key.get_type() == col_type_String || !num_identical_conditions || *num_identical_conditions < c_threshold_of_conditions_overwhelming_index))
             return false;
         return do_consume_condition(other);
+    }
+
+    constexpr static size_t c_threshold_of_conditions_overwhelming_index = 100;
+    bool num_conditions_may_need_combination_counts(size_t num_total_conditions)
+    {
+        return num_total_conditions >= c_threshold_of_conditions_overwhelming_index;
     }
 
     std::unique_ptr<ParentNode> m_child;
@@ -1163,16 +1175,17 @@ class FixedBytesNode<Equal, ObjectType, ArrayType> : public FixedBytesNodeBase<O
 public:
     using FixedBytesNodeBase<ObjectType, ArrayType>::FixedBytesNodeBase;
     using BaseType = FixedBytesNodeBase<ObjectType, ArrayType>;
+    using ThisType = FixedBytesNode<Equal, ObjectType, ArrayType>;
 
     void init(bool will_query_ranges) override
     {
         BaseType::init(will_query_ranges);
+        m_nb_needles = m_needles.size();
 
         if (!this->m_value_is_null) {
             m_optional_value = this->m_value;
         }
-
-        if (m_index_evaluator) {
+        if (m_index_evaluator && m_nb_needles == 0) {
             SearchIndex* index = BaseType::m_table->get_search_index(BaseType::m_condition_column_key);
             m_index_evaluator->init(index, m_optional_value);
             this->m_dT = 0;
@@ -1202,6 +1215,9 @@ public:
         size_t s = realm::npos;
 
         if (start < end) {
+            if (m_nb_needles) {
+                return find_first_haystack<22>(*this->m_leaf, m_needles, start, end);
+            }
             if (m_index_evaluator) {
                 return m_index_evaluator->do_search_index(this->m_cluster, start, end);
             }
@@ -1219,13 +1235,35 @@ public:
         return s;
     }
 
+    bool do_consume_condition(ParentNode& node) override
+    {
+        auto& other = static_cast<ThisType&>(node);
+        REALM_ASSERT(this->m_condition_column_key == other.m_condition_column_key);
+        REALM_ASSERT(other.m_needles.empty());
+        if (m_needles.empty()) {
+            m_needles.insert(this->m_value_is_null ? std::nullopt : std::make_optional(this->m_value));
+        }
+        m_needles.insert(other.m_value_is_null ? std::nullopt : std::make_optional(other.m_value));
+        return true;
+    }
+
     std::string describe(util::serializer::SerialisationState& state) const override
     {
         REALM_ASSERT(this->m_condition_column_key);
-        return state.describe_column(ParentNode::m_table, this->m_condition_column_key) + " " + Equal::description() +
-               " " +
-               (this->m_value_is_null ? util::serializer::print_value(realm::null())
-                                      : util::serializer::print_value(this->m_value));
+        std::string col_descr = state.describe_column(this->m_table, this->m_condition_column_key);
+        if (m_needles.empty()) {
+            return util::format("%1 %2 %3", col_descr, Equal::description(),
+                                (this->m_value_is_null ? util::serializer::print_value(realm::null())
+                                                       : util::serializer::print_value(this->m_value)));
+        }
+        std::string list_contents;
+        bool is_first = true;
+        for (auto it : m_needles) {
+            list_contents +=
+                util::format("%1%2", is_first ? "" : ", ", util::serializer::print_value(it)); // "it" may be null
+            is_first = false;
+        }
+        return util::format("%1 IN {%2}", col_descr, list_contents);
     }
 
     std::unique_ptr<ParentNode> clone() const override
@@ -1236,6 +1274,8 @@ public:
 protected:
     std::optional<ObjectType> m_optional_value;
     std::optional<IndexEvaluator> m_index_evaluator;
+    std::unordered_set<std::optional<ObjectType>> m_needles;
+    size_t m_nb_needles = 0;
 };
 
 
@@ -2101,17 +2141,65 @@ public:
     std::vector<std::unique_ptr<ParentNode>> m_conditions;
 
 private:
+    struct ConditionType {
+        ConditionType(const ParentNode* node)
+            : m_col(node->m_condition_column_key.value)
+            , m_type_hash(typeid(node).hash_code())
+        {
+        }
+        int64_t m_col;
+        size_t m_type_hash;
+        bool operator<(const ConditionType& other) const
+        {
+            return this->m_col < other.m_col && this->m_type_hash < other.m_type_hash;
+        }
+        bool operator!=(const ConditionType& other) const
+        {
+            return this->m_col != other.m_col || this->m_type_hash != other.m_type_hash;
+        }
+    };
+
     void combine_conditions(bool ignore_indexes)
     {
+        // Although ColKey is not unique per table, it is not important to consider
+        // the table when sorting here because
         std::sort(m_conditions.begin(), m_conditions.end(), [](auto& a, auto& b) {
-            return a->m_condition_column_key < b->m_condition_column_key;
+            return ConditionType(a.get()) < ConditionType(b.get());
         });
 
+        bool compute_condition_counts = num_conditions_may_need_combination_counts(m_conditions.size());
+        util::FlatMap<ConditionType, size_t> condition_type_counts;
+        if (compute_condition_counts) {
+            for (auto it = m_conditions.begin(); it != m_conditions.end();) {
+                // no need to try to combine anything other than simple nodes that
+                // filter directly on a top-level column since the only nodes that
+                // support combinations are string/int/uuid/oid <Equal> types
+                if (!(*it)->m_condition_column_key) {
+                    ++it;
+                    continue;
+                }
+                ConditionType cur_type((*it).get());
+                auto next = std::upper_bound(it, m_conditions.end(), cur_type,
+                                             [](const ConditionType& a, const std::unique_ptr<ParentNode>& b) {
+                                                 return a < ConditionType(b.get());
+                                             });
+                condition_type_counts[cur_type] = next - it;
+                it = next;
+            }
+        }
+
         auto prev = m_conditions.begin()->get();
+        std::optional<size_t> cur_type_count;
+        if (compute_condition_counts) {
+            cur_type_count = condition_type_counts[ConditionType{prev}];
+        }
         auto cond = [&](auto& node) {
-            if (prev->consume_condition(*node, ignore_indexes))
+            if (prev->consume_condition(*node, ignore_indexes, cur_type_count))
                 return true;
             prev = &*node;
+            if (compute_condition_counts) {
+                cur_type_count = condition_type_counts[ConditionType{prev}];
+            }
             return false;
         };
         m_conditions.erase(std::remove_if(m_conditions.begin() + 1, m_conditions.end(), cond), m_conditions.end());
@@ -2119,7 +2207,7 @@ private:
 
     // start index of the last find for each cond
     std::vector<size_t> m_start;
-    // last looked at index of the lasft find for each cond
+    // last looked at index of the last find for each cond
     // is a matching index if m_was_match is true
     std::vector<size_t> m_last;
     std::vector<bool> m_was_match;