Fixes https://github.com/realm/realm-dotnet/issues/3566

Added the ability for the UUID/ObjectId equality query node to consume conditions. 
As the number of conditions increases, the outer loop to individually check each condition increases more than checking each value against a hash set of condition values. This is true even if the UUID/ObjectId property is indexed.
![Screenshot 2024-04-11 at 11 12 59 AM](https://github.com/realm/realm-core/assets/2826060/dc64ddd2-f721-4e1a-9139-4130de387794)

Benchmark results as compared to the same benchmarks on master:

```
./test/benchmark-common-tasks/realm-benchmark-common-tasks visible ".*ChainedOr.*"
Results path: visible
Req runs:  145  QueryChainedOrStrings (MemOnly, EncryptionOff):     min   3.37ms (+0.05%)            max   3.72ms (+0.96%)            med   3.42ms (+0.02%)            avg   3.45ms (-0.20%)            stddev    73us (-12.42%)

Req runs:  366  QueryChainedOrStringsIndexed (MemOnly, EncryptionOff):     min   1.32ms (+0.37%)            max   1.54ms (+0.69%)            med   1.35ms (-0.22%)            avg   1.36ms (+0.06%)            stddev    32us (+4.05%)

Req runs:   17  QueryChainedOrStringsViewFilterPredicate (MemOnly, EncryptionOff):     min  28.14ms (+0.37%)            max  28.68ms (+0.61%)            med  28.35ms (+0.30%)            avg  28.33ms (+0.27%)            stddev   145us (+13.24%)

Req runs:   79  QueryNotChainedOrStrings (MemOnly, EncryptionOff):                     min   6.09ms (-0.30%)            max   6.82ms (+1.20%)            med   6.40ms (+1.75%)            avg   6.39ms (+1.19%)            stddev   188us (+30.81%)

Req runs:   79  QueryNotChainedOrStringsIndexed (MemOnly, EncryptionOff):              min   6.12ms (+0.36%)            max   6.74ms (-0.92%)            med   6.37ms (+1.39%)            avg   6.38ms (+0.88%)            stddev   145us (-16.76%)

Req runs:  538  QueryChainedOrInts (MemOnly, EncryptionOff):                           min    914us (+3.84%)            max   1.05ms (+1.50%)            med    924us (+3.49%)            avg    931us (+3.55%)            stddev    22us (+30.71%)

Req runs:  536  QueryChainedOrIntsIndexed (MemOnly, EncryptionOff):                    min    915us (-94.71%)           max   1.07ms (-94.03%)           med    926us (-94.74%)           avg    935us (-94.68%)           stddev    27us (-89.80%)

Req runs:  575  QueryChainedOrIntsCount (MemOnly, EncryptionOff):                      min    854us (+4.24%)            max   1.01ms (+4.14%)            med    862us (+4.24%)            avg    869us (+4.04%)            stddev    21us (-13.79%)

Req runs:  454  QueryChainedOrEquality_10<uuid><NonNullable><Indexed> (MemOnly, EncryptionOff):     min   1.07ms (+2.21%)            max   1.25ms (-3.20%)            med   1.08ms (+1.87%)            avg   1.09ms (+1.87%)            stddev    28us (-13.63%)

Req runs:  458  QueryChainedOrEquality_20<uuid><NonNullable><Indexed> (MemOnly, EncryptionOff):     min   1.07ms (+2.23%)            max   1.49ms (+19.73%)           med   1.08ms (+1.80%)            avg   1.10ms (+2.52%)            stddev    46us (+60.16%)

Req runs:  446  QueryChainedOrEquality_40<uuid><NonNullable><Indexed> (MemOnly, EncryptionOff):     min   1.07ms (+2.15%)            max   1.34ms (+2.75%)            med   1.09ms (+2.29%)            avg   1.11ms (+2.97%)            stddev    45us (+26.27%)

Req runs:  279  QueryChainedOrEquality_80<uuid><NonNullable><Indexed> (MemOnly, EncryptionOff):     min   1.74ms (+2.18%)            max   2.04ms (-0.21%)            med   1.78ms (+2.17%)            avg   1.80ms (+1.72%)            stddev    55us (-13.90%)

Req runs:  144  QueryChainedOrEquality_160<uuid><NonNullable><Indexed> (MemOnly, EncryptionOff):     min   3.37ms (-2.56%)            max   3.85ms (-2.75%)            med   3.44ms (-3.27%)            avg   3.47ms (-3.27%)            stddev    92us (-16.19%)

Req runs:  114  QueryChainedOrEquality_320<uuid><NonNullable><Indexed> (MemOnly, EncryptionOff):     min   4.28ms (-41.59%)           max   4.74ms (-42.17%)           med   4.39ms (-42.23%)           avg   4.41ms (-42.15%)           stddev    95us (-51.75%)

Req runs:   79  QueryChainedOrEquality_640<uuid><NonNullable><Indexed> (MemOnly, EncryptionOff):     min   6.05ms (-62.76%)           max   6.79ms (-62.15%)           med   6.20ms (-62.41%)           avg   6.25ms (-62.48%)           stddev   163us (-59.54%)

Req runs:   50  QueryChainedOrEquality_1280<uuid><NonNullable><Indexed> (MemOnly, EncryptionOff):     min   9.42ms (-71.36%)           max  10.21ms (-70.60%)           med   9.64ms (-71.01%)           avg   9.73ms (-70.83%)           stddev   248us (-53.94%)

Req runs:   29  QueryChainedOrEquality_2560<uuid><NonNullable><Indexed> (MemOnly, EncryptionOff):     min  16.25ms (-76.86%)           max  17.24ms (-76.31%)           med  16.40ms (-76.83%)           avg  16.46ms (-76.87%)           stddev   241us (-76.59%)

Req runs:   16  QueryChainedOrEquality_5120<uuid><NonNullable><Indexed> (MemOnly, EncryptionOff):     min  29.39ms (-81.73%)           max  30.97ms (-80.99%)           med  29.52ms (-81.81%)           avg  29.76ms (-81.64%)           stddev   486us (-44.23%)

Req runs:    8  QueryChainedOrEquality_10240<uuid><NonNullable><Indexed> (MemOnly, EncryptionOff):     min  54.94ms (-85.97%)           max  55.87ms (-85.97%)           med  55.38ms (-86.00%)           avg  55.36ms (-85.98%)           stddev   327us (-88.76%)

Req runs:    5  QueryChainedOrEquality_20480<uuid><NonNullable><Indexed> (MemOnly, EncryptionOff):     min 104.53ms (-89.76%)           max 109.13ms (-89.39%)           med 104.98ms (-89.77%)           avg 105.81ms (-89.68%)           stddev  1.93ms (-38.10%)

Req runs:   17  QueryChainedOrEquality_5000<uuid><NonNullable><Indexed> (MemOnly, EncryptionOff):      min  28.66ms (-81.49%)           max  30.39ms (-80.89%)           med  29.54ms (-81.12%)           avg  29.65ms (-81.08%)           stddev   489us (-72.39%)

Req runs:    5  QueryChainedOrEquality_5000<objectId><NonNullable><Indexed> (MemOnly, EncryptionOff):     min 192.98ms (-94.87%)           max 194.59ms (-94.90%)           med 193.14ms (-94.91%)           avg 193.42ms (-94.90%)           stddev   660us (-97.33%)

Req runs:   20  QueryChainedOrEquality_5000<uuid><NonNullable><NonIndexed> (MemOnly, EncryptionOff):      min  23.80ms (-97.72%)           max  25.82ms (-97.54%)           med  24.58ms (-97.65%)           avg  24.66ms (-97.65%)           stddev   487us (-55.60%)

Req runs:    5  QueryChainedOrEquality_5000<objectId><NonNullable><NonIndexed> (MemOnly, EncryptionOff):     min 138.12ms (-96.81%)           max 139.15ms (-96.80%)           med 138.65ms (-96.81%)           avg 138.56ms (-96.81%)           stddev   430us (-94.27%)

```

## ☑️ ToDos
* [x] 📝 Changelog update
* [x] 🚦 Tests (or not relevant)
* [x] C-API, if public C++ API changed
* [x] `bindgen/spec.yml`, if public C++ API changed